requirements.txt
streamlit==1.36.0
fpdf2==2.7.8
